// morphBlog/src/app/models/update-post.model.ts
export interface UpdatePost {
  title?: string | undefined;
  content?: string | undefined;
  summary?: string | undefined;
  urlHandle?: string | undefined;
  author?: string | undefined;
  visible?: string | undefined;
  publishDate?: Date | undefined;
  updateDate?: Date | undefined;
  featuredImage?: string | undefined;
}